requirements.txt
Django	1.11.5
Pygments	2.2.0
argparse	1.2.1
certifi	2017.7.27.1
chardet	3.0.4
django-rest-framework-mongoengine	3.3.1
djangorestframework	3.6.4
djangotoolbox	1.8.0
httpie	0.9.9
idna	2.6
mongoengine	0.13.0
numpy	1.13.3
pandas	0.20.3
pip	9.0.1
psycopg2	2.7.3.1
pymongo	3.5.1
python-dateutil	2.6.1
pytz	2017.2
requests	2.18.4
scikit-learn	0.19.0
scipy	0.19.1
setuptools	36.5.0
six	1.11.0
urllib3	1.22
wheel	0.30.0
wsgiref	0.1.2